requirements.txt
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pillow==10.4.0
requests==2.32.3
six==1.16.0
urllib3==1.26.20
yarl==1.11.1
psycopg2==2.9.9
textblob==0.18.0